// src/centerpanel/Flows/types/ObservationFormState.ts
export interface ObservationFormState {

  behaviorNarrative: string;

  imminentRiskType: string;
  riskClarifier: string;


  deescalationAttempts: string;
  nonPharmComfortsOffered: string;
  leastRestrictiveInsufficientWhy: string;


  observationLevel: string;
  observationRationale: string;
  monitoringPlanDetails: string;


  reassessmentFrequencyPlan: string;
  deescalationCriteria: string;
  handoffAndSupervision: string;
}